reorganize reader and writer modules
